Migrate calendar-user-address-set handling to calsched_support.c
