+++ Makefile.am
@@ -1153,6 +1153,8 @@ imap_libcyrus_imap_la_SOURCES += \
 	imap/ical_support.h \
 	imap/icu_wrap.h \
 	imap/icu_wrap.cpp \
+	imap/calsched_support.c \
+	imap/calsched_support.h \
 	imap/vcard_support.c \
 	imap/vcard_support.h \
 	imap/webdav_db.c \
