+++ imap/calsched_support.h
@@ -1,0 +1,72 @@
+/* calsched_support.h -- utility functions for dealing with calendar scheduling
+ *
+ * Copyright (c) 1994-2022 Carnegie Mellon University.  All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in
+ *    the documentation and/or other materials provided with the
+ *    distribution.
+ *
+ * 3. The name "Carnegie Mellon University" must not be used to
+ *    endorse or promote products derived from this software without
+ *    prior written permission. For permission or any legal
+ *    details, please contact
+ *      Carnegie Mellon University
+ *      Center for Technology Transfer and Enterprise Creation
+ *      4615 Forbes Avenue
+ *      Suite 302
+ *      Pittsburgh, PA  15213
+ *      (412) 268-7393, fax: (412) 268-7395
+ *      innovation@andrew.cmu.edu
+ *
+ * 4. Redistributions of any form whatsoever must retain the following
+ *    acknowledgment:
+ *    "This product includes software developed by Computing Services
+ *     at Carnegie Mellon University (http://www.cmu.edu/computing/)."
+ *
+ * CARNEGIE MELLON UNIVERSITY DISCLAIMS ALL WARRANTIES WITH REGARD TO
+ * THIS SOFTWARE, INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY
+ * AND FITNESS, IN NO EVENT SHALL CARNEGIE MELLON UNIVERSITY BE LIABLE
+ * FOR ANY SPECIAL, INDIRECT OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+ * WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN
+ * AN ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING
+ * OUT OF OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+ *
+ */
+
+#ifndef SCHED_UTIL_H
+#define SCHED_UTIL_H
+
+#include <libical/ical.h>
+
+#include "mailbox.h"
+#include "strarray.h"
+
+#define CALDAV_CALUSERADDR_INITIALIZER { STRARRAY_INITIALIZER, 0 }
+
+struct caldav_caluseraddr {
+    strarray_t uris;
+    int pref;
+};
+
+extern int caldav_caluseraddr_read(const char *mboxname,
+                                   const char *userid,
+                                   struct caldav_caluseraddr *addrs);
+
+extern int caldav_caluseraddr_write(struct mailbox *mbox,
+                                    const char *userid,
+                                    const struct caldav_caluseraddr *addrs);
+
+extern void caldav_caluseraddr_fini(struct caldav_caluseraddr *addr);
+
+extern void get_schedule_addresses(const char *mboxname,
+                                   const char *userid, strarray_t *addresses);
+
+#endif /* SCHED_UTIL_H */
