+++ imap/http_caldav_sched.h
@@ -59,6 +59,7 @@
 
 #endif /* WITH_DKIM */
 
+#include "calsched_support.h"
 #include "http_dav.h"
 #include "ical_support.h"
 #include "itip_support.h"
@@ -106,7 +107,6 @@ enum {
 extern unsigned config_allowsched;
 extern const char *ical_prodid;
 extern icaltimezone *utc_zone;
-extern struct strlist *cua_domains;
 extern icalarray *rscale_calendars;
 
 extern icalcomponent *busytime_query_local(struct transaction_t *txn,
@@ -142,7 +142,10 @@ extern xmlNodePtr xml_add_schedresponse(xmlNodePtr root, xmlNsPtr dav_ns,
                                         xmlChar *recipient, xmlChar *status);
 extern int caladdress_lookup(const char *addr, struct caldav_sched_param *param,
                              const strarray_t *schedule_addresses);
-extern void get_schedule_addresses(hdrcache_t req_hdrs, const char *mboxname,
-                                   const char *userid, strarray_t *addresses);
+
+extern void caldav_get_schedule_addresses(hdrcache_t req_hdrs,
+                                          const char *mboxname,
+                                          const char *userid,
+                                          strarray_t *addresses);
 
 #endif /* HTTP_CALDAV_SCHED_H */
